Simplify "unhandled entry" support with unique token. (now requires TS 2.7+)
