+++ package.json
@@ -34,10 +34,8 @@
         "pack": "run-p clean:pack build && npm pack",
         "jest": "jest",
         "jest:coverage": "npm run clean:coverage && jest --coverage",
-        "dtslint:v2_4_plus": "dtslint type_tests/v2_4_plus",
-        "dtslint:v2_6_plus": "dtslint type_tests/v2_6_plus",
         "dtslint:v2_7_plus": "dtslint type_tests/v2_7_plus",
-        "dtslint": "run-s clean:dist build:types dtslint:v2_4_plus dtslint:v2_6_plus dtslint:v2_7_plus",
+        "dtslint": "npm run dtslint:v2_7_plus",
         "test": "run-s compile lint prettier:test dtslint jest",
         "test:coverage": "run-s compile lint prettier:test dtslint jest:coverage",
         "build:travis": "run-p test:coverage build && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
@@ -63,7 +61,7 @@
         "typescript": "2.9.1"
     },
     "peerDependencies": {
-        "typescript": ">= 2.4.1"
+        "typescript": ">= 2.7.1"
     },
     "keywords": [
         "typescript",
