--- type_tests/v2_4_plus/tslint.json
@@ -1,8 +1,0 @@
-{
-    "extends": "dtslint/dtslint.json",
-    "rules": {
-        "newline-per-chained-call": false,
-        "no-useless-files": false,
-        "no-relative-import-in-test": false
-    }
-}
