+++ type_tests/v2_7_plus/mapString/enum-null.dtslint.ts
@@ -1,6 +1,10 @@
 import { mapString } from "../../../dist/types";
 
-type RGB = "r" | "g" | "b";
+enum RGB {
+    R = "r",
+    G = "g",
+    B = "b"
+}
 
 declare const rgb: RGB | null;
 
@@ -24,7 +28,7 @@ mapString(rgb).with({
 // $ExpectType number
 mapString(rgb).with({
     r: 10,
-    g: mapString.unhandled(),
+    g: mapString.unhandled,
     b: 30,
     handleNull: -1
 });
@@ -45,8 +49,8 @@ mapString(rgb).with({
     r: 10,
     g: 20,
     b: 30,
-    handleNull: mapString.unhandled(),
-    handleUnexpected: mapString.unhandled()
+    handleNull: mapString.unhandled,
+    handleUnexpected: mapString.unhandled
 });
 
 // Missing value handler causes error
