+++ type_tests/v2_7_plus/mapString/literal-null.dtslint.ts
@@ -2,28 +2,31 @@ import { mapString } from "../../../dist/types";
 
 type RGB = "r" | "g" | "b";
 
-declare const rgb: RGB;
+declare const rgb: RGB | null;
 
 // Return type is inferred
 // $ExpectType number
 mapString(rgb).with({
     r: 10,
     g: 20,
-    b: 30
+    b: 30,
+    handleNull: -1
 });
 // $ExpectType string
 mapString(rgb).with({
     r: "10",
     g: "20",
-    b: "30"
+    b: "30",
+    handleNull: "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 mapString(rgb).with({
     r: 10,
-    g: mapString.unhandled(),
-    b: 30
+    g: mapString.unhandled,
+    b: 30,
+    handleNull: -1
 });
 
 // handleUnexpected is allowed
@@ -32,6 +35,7 @@ mapString(rgb).with({
     r: 10,
     g: 20,
     b: 30,
+    handleNull: -1,
     handleUnexpected: -1
 });
 
@@ -41,14 +45,16 @@ mapString(rgb).with({
     r: 10,
     g: 20,
     b: 30,
-    handleUnexpected: mapString.unhandled()
+    handleNull: mapString.unhandled,
+    handleUnexpected: mapString.unhandled
 });
 
 // Missing value handler causes error
 // $ExpectError
 mapString(rgb).with({
     r: 10,
-    b: 30
+    b: 30,
+    handleNull: -1
 });
 
 // Unexpected value handler causes error
@@ -57,16 +63,16 @@ mapString(rgb).with({
     // $ExpectError
     oops: 42,
     g: 20,
-    b: 30
+    b: 30,
+    handleNull: -1
 });
 
-// Unnecessary null handler causes error
+// missing null handler causes error
+// $ExpectError
 mapString(rgb).with({
     r: 10,
     g: 20,
-    b: 30,
-    // $ExpectError
-    handleNull: -1
+    b: 30
 });
 
 // Unnecessary undefined handler causes error
@@ -74,6 +80,7 @@ mapString(rgb).with({
     r: 10,
     g: 20,
     b: 30,
+    handleNull: -1,
     // $ExpectError
     handleUndefined: -1
 });
