+++ type_tests/v2_7_plus/mapString/literal-undefined.dtslint.ts
@@ -1,33 +1,32 @@
 import { mapString } from "../../../dist/types";
 
-enum RGB {
-    R = "r",
-    G = "g",
-    B = "b"
-}
+type RGB = "r" | "g" | "b";
 
-declare const rgb: RGB;
+declare const rgb: RGB | undefined;
 
 // Return type is inferred
 // $ExpectType number
 mapString(rgb).with({
     r: 10,
     g: 20,
-    b: 30
+    b: 30,
+    handleUndefined: -1
 });
 // $ExpectType string
 mapString(rgb).with({
     r: "10",
     g: "20",
-    b: "30"
+    b: "30",
+    handleUndefined: "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 mapString(rgb).with({
     r: 10,
-    g: mapString.unhandled(),
-    b: 30
+    g: mapString.unhandled,
+    b: 30,
+    handleUndefined: -1
 });
 
 // handleUnexpected is allowed
@@ -36,6 +35,7 @@ mapString(rgb).with({
     r: 10,
     g: 20,
     b: 30,
+    handleUndefined: -1,
     handleUnexpected: -1
 });
 
@@ -45,14 +45,16 @@ mapString(rgb).with({
     r: 10,
     g: 20,
     b: 30,
-    handleUnexpected: mapString.unhandled()
+    handleUndefined: mapString.unhandled,
+    handleUnexpected: mapString.unhandled
 });
 
 // Missing value handler causes error
 // $ExpectError
 mapString(rgb).with({
     r: 10,
-    b: 30
+    b: 30,
+    handleUndefined: -1
 });
 
 // Unexpected value handler causes error
@@ -60,6 +62,15 @@ mapString(rgb).with({
     r: 10,
     // $ExpectError
     oops: 42,
+    g: 20,
+    b: 30,
+    handleUndefined: -1
+});
+
+// missing undefined handler causes error
+// $ExpectError
+mapString(rgb).with({
+    r: 10,
     g: 20,
     b: 30
 });
@@ -70,14 +81,6 @@ mapString(rgb).with({
     g: 20,
     b: 30,
     // $ExpectError
-    handleNull: -1
-});
-
-// Unnecessary undefined handler causes error
-mapString(rgb).with({
-    r: 10,
-    g: 20,
-    b: 30,
-    // $ExpectError
+    handleNull: -1,
     handleUndefined: -1
 });
