+++ type_tests/v2_7_plus/visitString/enum-null.dtslint.ts
@@ -6,7 +6,7 @@ enum RGB {
     B = "b"
 }
 
-declare const rgb: RGB;
+declare const rgb: RGB | null;
 
 // Test param types
 visitString(rgb).with({
@@ -22,8 +22,12 @@ visitString(rgb).with({
         // $ExpectType RGB.B
         value;
     },
+    handleNull: (value) => {
+        // $ExpectType null
+        value;
+    },
     handleUnexpected: (value) => {
-        // $ExpectType string | null | undefined
+        // $ExpectType string | undefined
         value;
     }
 });
@@ -32,7 +36,8 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
-    b: (value) => {}
+    b: (value) => {},
+    handleNull: (value) => {}
 });
 
 // Return type is inferred
@@ -40,21 +45,24 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
-    b: (value) => 30
+    b: (value) => 30,
+    handleNull: (value) => -1
 });
 // $ExpectType string
 visitString(rgb).with({
     r: (value) => "10",
     g: (value) => "20",
-    b: (value) => "30"
+    b: (value) => "30",
+    handleNull: (value) => "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 visitString(rgb).with({
     r: (value) => 10,
-    g: visitString.unhandled(),
-    b: (value) => 30
+    g: visitString.unhandled,
+    b: (value) => 30,
+    handleNull: (value) => -1
 });
 
 // special handlers can be unhandled
@@ -63,13 +71,23 @@ visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
     b: (value) => 30,
-    handleUnexpected: visitString.unhandled()
+    handleNull: visitString.unhandled,
+    handleUnexpected: visitString.unhandled
 });
 
 // Missing value handler causes error
 // $ExpectError
 visitString(rgb).with({
     r: (value) => {},
+    b: (value) => {},
+    handleNull: (value) => {}
+});
+
+// Missing null handler causes error
+// $ExpectError
+visitString(rgb).with({
+    r: (value) => {},
+    g: (value) => {},
     b: (value) => {}
 });
 
@@ -79,15 +97,7 @@ visitString(rgb).with({
     // $ExpectError
     oops: (value) => {},
     g: (value) => {},
-    b: (value) => {}
-});
-
-// Unnecessary null handler causes error
-visitString(rgb).with({
-    r: (value) => {},
-    g: (value) => {},
     b: (value) => {},
-    // $ExpectError
     handleNull: (value) => {}
 });
 
@@ -96,6 +106,7 @@ visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
     b: (value) => {},
+    handleNull: (value) => {},
     // $ExpectError
     handleUndefined: (value) => {}
 });
