+++ type_tests/v2_7_plus/visitString/literal-both.dtslint.ts
@@ -1,33 +1,33 @@
 import { visitString } from "../../../dist/types";
 
-enum RGB {
-    R = "r",
-    G = "g",
-    B = "b"
-}
+type RGB = "r" | "g" | "b";
 
-declare const rgb: RGB | null;
+declare const rgb: RGB | null | undefined;
 
 // Test param types
 visitString(rgb).with({
     r: (value) => {
-        // $ExpectType RGB.R
+        // $ExpectType "r"
         value;
     },
     g: (value) => {
-        // $ExpectType RGB.G
+        // $ExpectType "g"
         value;
     },
     b: (value) => {
-        // $ExpectType RGB.B
+        // $ExpectType "b"
         value;
     },
     handleNull: (value) => {
         // $ExpectType null
         value;
     },
+    handleUndefined: (value) => {
+        // $ExpectType undefined
+        value;
+    },
     handleUnexpected: (value) => {
-        // $ExpectType string | undefined
+        // $ExpectType string
         value;
     }
 });
@@ -37,7 +37,8 @@ visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
     b: (value) => {},
-    handleNull: (value) => {}
+    handleNull: (value) => {},
+    handleUndefined: (value) => {}
 });
 
 // Return type is inferred
@@ -46,23 +47,26 @@ visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
     b: (value) => 30,
-    handleNull: (value) => -1
+    handleNull: (value) => -1,
+    handleUndefined: (value) => -1
 });
 // $ExpectType string
 visitString(rgb).with({
     r: (value) => "10",
     g: (value) => "20",
     b: (value) => "30",
-    handleNull: (value) => "-1"
+    handleNull: (value) => "-1",
+    handleUndefined: (value) => "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 visitString(rgb).with({
     r: (value) => 10,
-    g: visitString.unhandled(),
+    g: visitString.unhandled,
     b: (value) => 30,
-    handleNull: (value) => -1
+    handleNull: (value) => -1,
+    handleUndefined: (value) => -1
 });
 
 // special handlers can be unhandled
@@ -71,8 +75,9 @@ visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
     b: (value) => 30,
-    handleNull: visitString.unhandled(),
-    handleUnexpected: visitString.unhandled()
+    handleNull: visitString.unhandled,
+    handleUndefined: visitString.unhandled,
+    handleUnexpected: visitString.unhandled
 });
 
 // Missing value handler causes error
@@ -80,7 +85,8 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => {},
     b: (value) => {},
-    handleNull: (value) => {}
+    handleNull: (value) => {},
+    handleUndefined: (value) => {}
 });
 
 // Missing null handler causes error
@@ -88,7 +94,17 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
-    b: (value) => {}
+    b: (value) => {},
+    handleUndefined: (value) => {}
+});
+
+// Missing undefined handler causes error
+// $ExpectError
+visitString(rgb).with({
+    r: (value) => {},
+    g: (value) => {},
+    b: (value) => {},
+    handleNull: (value) => {}
 });
 
 // Unexpected value handler causes error
@@ -98,15 +114,6 @@ visitString(rgb).with({
     oops: (value) => {},
     g: (value) => {},
     b: (value) => {},
-    handleNull: (value) => {}
-});
-
-// Unnecessary undefined handler causes error
-visitString(rgb).with({
-    r: (value) => {},
-    g: (value) => {},
-    b: (value) => {},
     handleNull: (value) => {},
-    // $ExpectError
     handleUndefined: (value) => {}
 });
