+++ type_tests/v2_7_plus/visitString/literal-null.dtslint.ts
@@ -2,7 +2,7 @@ import { visitString } from "../../../dist/types";
 
 type RGB = "r" | "g" | "b";
 
-declare const rgb: RGB | null | undefined;
+declare const rgb: RGB | null;
 
 // Test param types
 visitString(rgb).with({
@@ -22,12 +22,8 @@ visitString(rgb).with({
         // $ExpectType null
         value;
     },
-    handleUndefined: (value) => {
-        // $ExpectType undefined
-        value;
-    },
     handleUnexpected: (value) => {
-        // $ExpectType string
+        // $ExpectType string | undefined
         value;
     }
 });
@@ -37,8 +33,7 @@ visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
     b: (value) => {},
-    handleNull: (value) => {},
-    handleUndefined: (value) => {}
+    handleNull: (value) => {}
 });
 
 // Return type is inferred
@@ -47,26 +42,23 @@ visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
     b: (value) => 30,
-    handleNull: (value) => -1,
-    handleUndefined: (value) => -1
+    handleNull: (value) => -1
 });
 // $ExpectType string
 visitString(rgb).with({
     r: (value) => "10",
     g: (value) => "20",
     b: (value) => "30",
-    handleNull: (value) => "-1",
-    handleUndefined: (value) => "-1"
+    handleNull: (value) => "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 visitString(rgb).with({
     r: (value) => 10,
-    g: visitString.unhandled(),
+    g: visitString.unhandled,
     b: (value) => 30,
-    handleNull: (value) => -1,
-    handleUndefined: (value) => -1
+    handleNull: (value) => -1
 });
 
 // special handlers can be unhandled
@@ -75,9 +67,8 @@ visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
     b: (value) => 30,
-    handleNull: visitString.unhandled(),
-    handleUndefined: visitString.unhandled(),
-    handleUnexpected: visitString.unhandled()
+    handleNull: visitString.unhandled,
+    handleUnexpected: visitString.unhandled
 });
 
 // Missing value handler causes error
@@ -85,8 +76,7 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => {},
     b: (value) => {},
-    handleNull: (value) => {},
-    handleUndefined: (value) => {}
+    handleNull: (value) => {}
 });
 
 // Missing null handler causes error
@@ -94,17 +84,7 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
-    b: (value) => {},
-    handleUndefined: (value) => {}
-});
-
-// Missing undefined handler causes error
-// $ExpectError
-visitString(rgb).with({
-    r: (value) => {},
-    g: (value) => {},
-    b: (value) => {},
-    handleNull: (value) => {}
+    b: (value) => {}
 });
 
 // Unexpected value handler causes error
@@ -114,6 +94,15 @@ visitString(rgb).with({
     oops: (value) => {},
     g: (value) => {},
     b: (value) => {},
+    handleNull: (value) => {}
+});
+
+// Unnecessary undefined handler causes error
+visitString(rgb).with({
+    r: (value) => {},
+    g: (value) => {},
+    b: (value) => {},
     handleNull: (value) => {},
+    // $ExpectError
     handleUndefined: (value) => {}
 });
