+++ type_tests/v2_7_plus/visitString/literal-undefined.dtslint.ts
@@ -2,7 +2,7 @@ import { visitString } from "../../../dist/types";
 
 type RGB = "r" | "g" | "b";
 
-declare const rgb: RGB;
+declare const rgb: RGB | undefined;
 
 // Test param types
 visitString(rgb).with({
@@ -18,8 +18,12 @@ visitString(rgb).with({
         // $ExpectType "b"
         value;
     },
+    handleUndefined: (value) => {
+        // $ExpectType undefined
+        value;
+    },
     handleUnexpected: (value) => {
-        // $ExpectType string | null | undefined
+        // $ExpectType string | null
         value;
     }
 });
@@ -28,7 +32,8 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
-    b: (value) => {}
+    b: (value) => {},
+    handleUndefined: (value) => {}
 });
 
 // Return type is inferred
@@ -36,21 +41,24 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
-    b: (value) => 30
+    b: (value) => 30,
+    handleUndefined: (value) => -1
 });
 // $ExpectType string
 visitString(rgb).with({
     r: (value) => "10",
     g: (value) => "20",
-    b: (value) => "30"
+    b: (value) => "30",
+    handleUndefined: (value) => "-1"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 visitString(rgb).with({
     r: (value) => 10,
-    g: visitString.unhandled(),
-    b: (value) => 30
+    g: visitString.unhandled,
+    b: (value) => 30,
+    handleUndefined: (value) => -1
 });
 
 // special handlers can be unhandled
@@ -59,13 +67,23 @@ visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
     b: (value) => 30,
-    handleUnexpected: visitString.unhandled()
+    handleUndefined: visitString.unhandled,
+    handleUnexpected: visitString.unhandled
 });
 
 // Missing value handler causes error
 // $ExpectError
 visitString(rgb).with({
     r: (value) => {},
+    b: (value) => {},
+    handleUndefined: (value) => {}
+});
+
+// Missing undefined handler causes error
+// $ExpectError
+visitString(rgb).with({
+    r: (value) => {},
+    g: (value) => {},
     b: (value) => {}
 });
 
@@ -75,7 +93,8 @@ visitString(rgb).with({
     // $ExpectError
     oops: (value) => {},
     g: (value) => {},
-    b: (value) => {}
+    b: (value) => {},
+    handleUndefined: (value) => {}
 });
 
 // Unnecessary null handler causes error
@@ -84,14 +103,6 @@ visitString(rgb).with({
     g: (value) => {},
     b: (value) => {},
     // $ExpectError
-    handleNull: (value) => {}
-});
-
-// Unnecessary undefined handler causes error
-visitString(rgb).with({
-    r: (value) => {},
-    g: (value) => {},
-    b: (value) => {},
-    // $ExpectError
+    handleNull: (value) => {},
     handleUndefined: (value) => {}
 });
