+++ type_tests/v2_7_plus/visitString/literal.dtslint.ts
@@ -2,7 +2,7 @@ import { visitString } from "../../../dist/types";
 
 type RGB = "r" | "g" | "b";
 
-declare const rgb: RGB | null;
+declare const rgb: RGB;
 
 // Test param types
 visitString(rgb).with({
@@ -18,12 +18,8 @@ visitString(rgb).with({
         // $ExpectType "b"
         value;
     },
-    handleNull: (value) => {
-        // $ExpectType null
-        value;
-    },
     handleUnexpected: (value) => {
-        // $ExpectType string | undefined
+        // $ExpectType string | null | undefined
         value;
     }
 });
@@ -32,8 +28,7 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
-    b: (value) => {},
-    handleNull: (value) => {}
+    b: (value) => {}
 });
 
 // Return type is inferred
@@ -41,24 +36,21 @@ visitString(rgb).with({
 visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
-    b: (value) => 30,
-    handleNull: (value) => -1
+    b: (value) => 30
 });
 // $ExpectType string
 visitString(rgb).with({
     r: (value) => "10",
     g: (value) => "20",
-    b: (value) => "30",
-    handleNull: (value) => "-1"
+    b: (value) => "30"
 });
 
 // Return type is inferred when "unhandled" entries exist
 // $ExpectType number
 visitString(rgb).with({
     r: (value) => 10,
-    g: visitString.unhandled(),
-    b: (value) => 30,
-    handleNull: (value) => -1
+    g: visitString.unhandled,
+    b: (value) => 30
 });
 
 // special handlers can be unhandled
@@ -67,23 +59,13 @@ visitString(rgb).with({
     r: (value) => 10,
     g: (value) => 20,
     b: (value) => 30,
-    handleNull: visitString.unhandled(),
-    handleUnexpected: visitString.unhandled()
+    handleUnexpected: visitString.unhandled
 });
 
 // Missing value handler causes error
 // $ExpectError
 visitString(rgb).with({
     r: (value) => {},
-    b: (value) => {},
-    handleNull: (value) => {}
-});
-
-// Missing null handler causes error
-// $ExpectError
-visitString(rgb).with({
-    r: (value) => {},
-    g: (value) => {},
     b: (value) => {}
 });
 
@@ -93,7 +75,15 @@ visitString(rgb).with({
     // $ExpectError
     oops: (value) => {},
     g: (value) => {},
+    b: (value) => {}
+});
+
+// Unnecessary null handler causes error
+visitString(rgb).with({
+    r: (value) => {},
+    g: (value) => {},
     b: (value) => {},
+    // $ExpectError
     handleNull: (value) => {}
 });
 
@@ -102,7 +92,6 @@ visitString(rgb).with({
     r: (value) => {},
     g: (value) => {},
     b: (value) => {},
-    handleNull: (value) => {},
     // $ExpectError
     handleUndefined: (value) => {}
 });
